fix(rpg-12): rewrite Load Capacity as exact-integer formulas, drop the Strength offset
Replaces the rpg-11 kLoadCapacityStrengthOffset correction with a new
design: kMaxCapacityDivisor raised from 25 to 150 so MaxCapacityKg
works directly off the baseline-60 Strength (no offset needed).
LightLoadKg/HeavyLoadKg switch from 0.3/0.7 fractions to exact
floor(MaxCapacityKg/3) and floor(MaxCapacityKg*2/3). All four Load
Capacity methods (and LoadCapacityResponse) now return int instead of
double, matching the design document's own integer arithmetic. At
human defaults: MaxCapacityKg=84, LightLoadKg=28, HeavyLoadKg=56,
DragCapacityKg=203 (DragCapacityKg's own formula is unchanged).

+++ src/main/java/com/keynor/rpg/application/dto/LoadCapacityResponse.java
@@ -2,8 +2,7 @@ package com.keynor.rpg.application.dto;
 
 import com.keynor.rpg.domain.model.PlayableCharacter;
 
-public record LoadCapacityResponse(double lightLoadKg, double heavyLoadKg, double maxCapacityKg,
-                                    double dragCapacityKg) {
+public record LoadCapacityResponse(int lightLoadKg, int heavyLoadKg, int maxCapacityKg, int dragCapacityKg) {
 
     public static LoadCapacityResponse from(PlayableCharacter character) {
         return new LoadCapacityResponse(character.getLightLoadKg(), character.getHeavyLoadKg(),
